Test Cleanup Fixes #25
